implemented platform utils string to version with tests
GitOrigin-RevId: 5765148464db7df267d651bcabc27f343c94ac96

+++ v11/larkin/environment/system/platform/platform_info.h
@@ -29,9 +29,9 @@ enum OperatingSystem {
 // Platform version numbers following a major, minor,
 // and patch number system.
 struct Version {
-    int major_version;
-    int minor_version;
-    int patch_version;
+    uint16_t major_version;
+    uint16_t minor_version;
+    uint16_t patch_version;
 };
 
 }  // namespace sys
